Fit the bigger register in memory, and fix the score that made light mode lie
The register grew thirteenfold and the hosted deployment started returning 502
on every request. Measured, the process held 773 MB against a 512 MB limit, and
166 MB of that was data the bigger register brought with it.

Three reductions, together 166 MB -> 123 MB:

  * embeddings stored at float16. They are L2-normalised, so every component is
    within [-1, 1] where half precision carries about three decimal digits —
    finer than the gaps between ranked cosine scores. Ranking is unchanged on
    the golden set. 43 MB -> 21 MB, on disk and in memory.
  * the corpus keeps twelve columns instead of seventeen. Holding fields no
    caller reads cost 39 MB of dictionaries.
  * the tokenised documents are freed once BM25 has built its tables, which it
    never reads again. 31 MB.

Measuring light mode as a fallback exposed something worse than a memory
problem. Without a cross-encoder the fused rank stands in for a score, and it
was normalised against the best candidate actually found — so the leader scored
1.0 whatever it was. 'banana republic' came back at score 1.0 and clear_match.
The confidence gate, which is the property this system is built around, could
not fire at all.

Reciprocal rank fusion is bounded: a candidate placed first by every ranking
scores len(rankings)/(RRF_K + 1). Calibrating against that ceiling rather than
against the observed top restores the gate — nonsense now scores 0.33 and
abstains, real queries score above 0.98.

Worth recording: on a 27,687-standard register the cross-encoder now costs
rank-1 accuracy rather than adding it. Light mode scores 59/71 against full
mode's 56/71, with both at 66/71 for recall@10. ms-marco was trained on web
passages, and fusion over two retrievers appears to suit standards titles
better. Not acted on yet — it deserves a proper comparison, not a switch flipped
at four in the morning.

28 of 28 adversarial checks pass.

+++ build_embeddings.py
@@ -18,7 +18,11 @@ def main():
     print(f"Embedding {len(texts)} standards...")
     embeddings = model.encode(texts, show_progress_bar=True, normalize_embeddings=True)
 
-    np.save(EMBEDDINGS_PATH, embeddings)
+    # Stored at half precision. The vectors are L2-normalised so every component
+    # is within [-1, 1], where float16 carries about three decimal digits — finer
+    # than the gaps between ranked cosine scores, and half the file and half the
+    # memory on a host that has 512 MB for everything.
+    np.save(EMBEDDINGS_PATH, embeddings.astype(np.float16))
     df[["IS Number"]].to_csv(IS_NUMBERS_PATH, index=False)
 
     print(f"\nSaved embeddings: {EMBEDDINGS_PATH} shape={embeddings.shape}")
